v4.2.343
FUNCTIONALITY:
-Check for updates
-Script has 3 channels: Stable (v), Beta (b), and DEV (d)
-Speed up antihang (detects when command ends immediately)

SETTINGS:
-new: check4update (1||0)

When I first made this script, It was just something small I could use to help with frequent loss of connection in an apartment complex where everybody and their dog had routers and wireless devices and everything interfered with each other. My Network_Resetter worked fine for the occasional loss of connection, but was nowhere near quick enough to keep up with the frequent signal interference at that location.

The first version was something I quickly typed up, only took a couple hours.I believed I was done with it. Then about a week later I made some changes, rewrote some things, made more of the script automatic. Again, I thought it was done. This has continued for every new version I've released of this script. Every time I think I'm done, I think of something else to change or add or fix or automate...whatever the case may be.

Had I known how many versions of this I'd release, I would have looked into implementing an update check much sooner. I had only considered automatic updates (issue 5), but that posed to many impracticalities, the biggest being the false positive detection by certain antivirus.

Looking into it again, I have found methods that are suitable for the purposes of checking for updates; after scanning the script online there seem to be no false positives. It will be up to the user to choose when and if to update, but the script will (by default) let them know if an update is available. The script will check about 5 to 10 minutes after it starts, and once about every 24 hours after that.

There are 3 methods the script now uses to check for updates. The first reads the remote text file via Powershell 2.0. This method is tried first since it does not require creating an external file. If this fails, the 2nd method is BITSADMIN. This requires write access as it downloads the remote text file and saves it temporarily. The third option uses a temporary VBS script to download the remote text file. The last option has the most compatibility, working on systems since Windows 2000 and still works on the latest systems (Win2012 R2 at the time this was written)