v5.0.354
-Incremented major version number
-Renamed 'nwcr_simple.cmd' to 'quick-net-fix.cmd'

With the release v4.3.353 came many new features and functionality. Among the changes were notable ones like Update checks,  indirect execution of commands that risk hanging indefinitely, and more compatibility checks for the availability of commonly used commands. However, many of the changes sacrificed speed and efficiency.

With a less efficient script, the time it takes to complete certain tasks may not be too much longer on fast computers, but on slower ones tasks that took seconds could take minutes.

With this in mind, one of the primary goals now is to strip the script down to bare essentials, rewrite what is left to make it streamlined and efficient, then re-add the various features piece by piece. Features that are not necessary will not be re-added. Speed and efficiency will be top priorities.

The rename was overdue, and the next version will put the 'Quick' back in 'Quick Net Fix'