v5.0.364
FUNCTIONALITY:
-Quickly verify validity of curRTR after each connectivity test
-If curRTR is invalid, check for existence of valid router; if valid router exists call ":chkRA"; if it does not exist set curRTR to fallback router address

NOTE: The above functionality changes ensure that the fallback router addresses are only used as a last resort. If there is a "blip" in the connection, it will not immediately switch to the fallback address. If the router address has changed, the script will detect it immediately *without the need for a successful connectivity test*. This makes this script ideal to keep computers connected to a network with or without internet access. On a network without internet access, by the time the script uses the fallback addresses, there is no other address available that would result in a successful connectivity test.

CODE CLEANUP:
-Significantly speed up ":getNETINFO" (Verify Adapter/Router) by parsing only the lines with specific key words
-Clear all variables with the prefix 'net_' before parsing IPCONFIG
-Minor code compression