v5.0.369
SETTINGS:
-New: TME_c4uwait (wait at least specified time between update checks)

New Type of setting: TME. The value for this type of setting is time in the format [seconds minutes hours days months years]. All but [seconds] can be omitted. For example, a period of 10 minutes is written as [0 10]. 2 months and 4 hours is written as [0 0 4 0 2]. Values cannot exceed: [59 59 23 31 11 2000000000]. Brackets are used here for readability purposes only; they are not a part of the setting format.

FUNCTIONALITY:
-Update Check delay is now configurable. While it is still an estimate (it will be off by a few seconds in most cases), it is much more precise (previously it had the potential of being off by many hours or even days). In addition to the more precise measurements, it will now take into account periods in which the script is paused (such as when the computer is in stand by / sleep mode).
-Leftover temp files (if any) are now cleaned once every hour and 10 minutes (as opposed to once anywhere from 15 minutes to an hour).